+++ library/constants.py
@@ -1,0 +1,6 @@
+RATE = 44100
+BPM = 80
+
+__all__ = [
+    'RATE', 'BPM',
+]
